Improve error handling and code consistency
- Added error handling for memory allocation and file operations in multiple functions.
- Updated functions to handle NULL parameters and improve logic flow.
- Simplified comparison logic in `find_car` and improved type safety in `turn_string_to_us_int`.
- Enhanced error management with new functions in `error_handle.c`.
- Renamed `printDate` to `print_Date` for consistency and updated related functions.
- Improved user prompts and logging in the `main` function and other areas.

+++ app/inputer.c
@@ -269,7 +269,7 @@ unsigned* turn_string_to_us_int(const char* str)
 			return NULL; // Return NULL if invalid character found
 		}
 		int temp = (str[i] - '0');
-		sum = sum+ pow(10,u)*(str[i] - '0');
+		sum = sum+ (int)(pow(10,u)*(str[i] - '0'));
 		u++; i--;
 	}
 	p = &sum;
